Add support for MaxIdleTimeClosed stat in go version 1.15+
