+++ go.mod
@@ -2,4 +2,8 @@ module github.com/dlmiddlecote/sqlstats
 
 go 1.11
 
-require github.com/prometheus/client_golang v1.3.0
+require (
+	github.com/mattn/go-sqlite3 v1.14.6
+	github.com/prometheus/client_golang v1.3.0
+	github.com/stretchr/testify v1.3.0
+)
